Added check for size

+++ 0x1E-search_algorithms/0-linear.c
@@ -8,11 +8,11 @@
  */
 int linear_search(int *array, size_t size, int value)
 {
-	int i;
+	size_t i;
 
 	if (array == NULL || size == 0)
 		return (-1);
-	for (i = 0; i < int(size); i++)
+	for (i = 0; i < size; i++)
 	{
 		printf("Value checked at array[%lu] = [%d]\n", i, array[i]);
 		if (array[i] == value)
